testing views: 'show-cards' and 'deal-cards' populated database

+++ src/main/java/com/blackjack/project/card/CardRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.repository.Repository;
 
 import java.util.List;
 
-public interface CardRepository extends Repository<Card, Long> {
+public interface CardRepository extends Repository<Card, Integer> {
 
     List<Card> findAll();
